split event into 2 separate events for add and removal of status.
